// proyectoFinal/src/model/types/student.ts
export interface Student {
    id: number;
    name: number;
    firstSurname: string;
    secondSurname: string;
    personalEmailAddress: string;
    activaEmailAddress: string;
    description: string;
    city: string;
    phoneNumber: string;
    zipCode: string;
    avatar: any;
    cv: any;
}